+++ src/management-system/src/backend/server/iam/rest-api/auth0/user.js
@@ -4,8 +4,8 @@ import { validateUser } from '../../middleware/inputValidations.js';
 import { ensureCleanRoleMappings } from '../../utils/roleMappings.js';
 import { isAllowed, isAuthenticated } from '../../middleware/authorization';
 import { config } from '../../../iam/utils/config.js';
-import { toCaslResource } from '../../authorization/caslRules';
 import Ability from '../../../../../../../management-system-v2/lib/ability/abilityHelper';
+import { toCaslResource } from '../../../../../../../management-system-v2/lib/ability/caslAbility';
 
 // some userinfo endpoints are not returning standardized fields, so we have to make a standard for the MS
 const standardizeUser = (user) => {
