+++ src/management-system/src/backend/server/iam/rest-api/keycloak/user.js
@@ -3,8 +3,8 @@ import requestResource from '../index.js';
 import { validateUser } from '../../middleware/inputValidations.js';
 import { ensureCleanRoleMappings } from '../../utils/roleMappings.js';
 import { isAllowed, isAuthenticated } from '../../middleware/authorization';
-import { toCaslResource } from '../../authorization/caslRules';
 import Ability from '../../../../../../../management-system-v2/lib/ability/abilityHelper';
+import { toCaslResource } from '../../../../../../../management-system-v2/lib/ability/caslAbility';
 
 const userRouter = express.Router();
