+++ src/management-system/src/backend/server/rest-api/process.js
@@ -18,7 +18,7 @@ import express from 'express';
 import { isAllowed } from '../iam/middleware/authorization.ts';
 import logger from '../../shared-electron-server/logging.js';
 import Ability from '../../../../../management-system-v2/lib/ability/abilityHelper';
-import { toCaslResource } from '../iam/authorization/caslRules';
+import { toCaslResource } from '../../../../../management-system-v2/lib/ability/caslAbility';
 
 const processRouter = express.Router();
