+++ src/management-system/tests/unit/specs/backend/authorization/iam/users.spec.js
@@ -1,10 +1,10 @@
 const { users, abilitiesBuilt } = require('../authorization');
 const {
-  toCaslResource,
-} = require('../../../../../../src/backend/server/iam/authorization/caslRules');
-const {
   permissionNumberToIdentifiers,
 } = require('../../../../../../src/backend/server/iam/authorization/permissionHelpers');
+const {
+  toCaslResource,
+} = require('../../../../../../../management-system-v2/lib/ability/caslAbility');
 
 let adminAbility,
   all_user_permissionsAbility,
