+++ src/management-system/tests/unit/specs/backend/authorization/machines.spec.js
@@ -1,8 +1,10 @@
 const { abilitiesBuilt, buildAbility } = require('./authorization');
-const { toCaslResource } = require('../../../../../src/backend/server/iam/authorization/caslRules');
 const {
   permissionNumberToIdentifiers,
 } = require('../../../../../src/backend/server/iam/authorization/permissionHelpers');
+const {
+  toCaslResource,
+} = require('../../../../../../management-system-v2/lib/ability/caslAbility');
 
 let adminAbility, all_role_permissionsAbility, unauthenticatedAbility;
