+++ src/management-system/tests/unit/specs/backend/authorization/shares/getShare.spec.js
@@ -1,13 +1,13 @@
 const { processes, shares, abilitiesBuilt, buildAbility } = require('../authorization');
 const {
-  toCaslResource,
-} = require('../../../../../../src/backend/server/iam/authorization/caslRules');
-const {
   permissionNumberToIdentifiers,
 } = require('../../../../../../src/backend/server/iam/authorization/permissionHelpers');
 const {
   getShares,
 } = require('../../../../../../src/backend/shared-electron-server/data/iam/shares');
+const {
+  toCaslResource,
+} = require('../../../../../../../management-system-v2/lib/ability/caslAbility');
 
 let adminAbility, unauthenticatedAbility;
